feat(008): لوحة تصحيح المقالي — ونصفُ الدرجة الذي كان يختفي (US5)
المصحّح يرى المنتظر، يوزّع الدرجة على معايير Rubric مع تعليقٍ لكلّ معيار،
فيُحتسب المجموع وتُبلَّغ الطالب وتُعاد أهليّة الشهادة.

⚠️ نصفُ درجةٍ كان يختفي بعد أن يمرّ من كلّ فحص. exam_answers.points كان
unsignedSmallInteger: معياران بـ٢٫٥ يجتازان «المجموع ≤ درجة السؤال» ثمّ يفقدان
نصفَيهما عند الكتابة — بعد كلّ تحقّق، وبمقدار ما تدقّق المدرّس. صار عشرياً
بإشارة: الحساب بلا إشارة يرمي ERROR 1690 على MySQL وحدها، وSQLite التي تعمل
عليها الاختبارات لا تعرف الفيضان أصلاً.

⚠️ ولا قيد فريد يحرس من مصحّحَين. الشكل الواضح
unique(answer_id, rubric_criterion_id, revision_of) عمودان منه قابلان للإفراغ:
مقاليٌّ بلا معايير يكتب NULL فيهما، وNULL لا يصطدم بـNULL — فينجح الاثنان
بصمت، في الشكل الذي تُصحَّح به أكثر المقاليات. الحارس مطالبةٌ ذرّية على
exam_answers.graded_at، وgrading_records جدولٌ للإلحاق فقط يمنعه النموذج نفسه.

⚠️ وبابُ التعديل كان يتخطّى الطابور، والضرر يعيش بعده. ورقةٌ فيها مقاليّان:
يصحّح الأوّل، يلمح خطأً فيعدّله — نقرةٌ واحدة — فتُختَم الورقة كاملةً والثاني
لم يفتحه أحد: ExamPassed، الإشعار، سلسلة الشهادة. والأسوأ لاحقاً: الإغلاق
يتوقّف عند «لم تعد منتظرة»، فدرجةُ الثاني تُكتب في صفّها ولا تدخل المجموع
أبداً. ReviseGrade لا يُنهي ورقةً ما زالت pending_grading.

- صلاحيتان لا واحدة: التصحيح عملٌ يفوَّض، وتعديلُ درجةٍ قيلت للطالب قرارُ
  المدرّس بنقضه. صلاحيةٌ واحدة تُعيد للمساعد كتابة درجات زميله.
- إخفاء الهوية في الـResource لا في الشاشة (FR-033)، والمفتاح غائبٌ لا فارغ،
  وإطفاؤه يُسجَّل في activity_log.
- AttemptPolicy لم يعد ATTEMPTS_VIEW_ALL وحدها (NFR-001أ): تسجيلٌ نشط، أو
  محاولةٌ وقعت داخل تسجيل — والفرع الثاني يمنع ورقةَ آخر يومٍ في الفصل من
  التعلّق بلا مصحّح.
- إشعارٌ مستقلّ لـ«تسلّمنا ورقتك»: الدرجةُ الآلية وحدها على ورقةٍ فيها مقاليّ
  تقول لطالبٍ كتب إجاباتٍ ممتازة إنه رسب.
- ميزانية استعلامات مثبَّتة بـ١٥ للطابور: خمس أوراق فيكسشر، وأوّل يومٍ في
  الفصل خمسمئة.

وأُصلح صنفا لونٍ ميّتان من US4: bg-success-soft وbg-danger-soft ليسا رمزين في
@theme، وTailwind لا يُصدر قاعدةً لرمزٍ غير موجود — فكانت دائرةُ الصواب والخطأ
بلا تعبئة.

pest 1268/1268 · playwright e2e/grading 16/16 · pint · phpstan L8 · tsc

+++ backend/app/Modules/Assessments/Actions/FinalizeAttempt.php
@@ -36,7 +36,10 @@ class FinalizeAttempt extends Action
             $items = $attempt->items()->get();
 
             $totalPoints = (int) $items->sum('points');
-            $earnedPoints = (int) $answers->sum('points');
+            // ⚠️ FLOAT, NOT INT. A rubric awards halves, and an `(int)` here
+            // rounds the whole paper down once per essay — after the mark scheme
+            // has already been validated as adding up.
+            $earnedPoints = (float) $answers->sum('points');
 
             $maxScore = max($totalPoints, 1);
             $scorePct = round(($earnedPoints / $maxScore) * 100, 2);
